refactor: Explode responsability
